Add sandbox forking, state persistence, and replace Docker with Podman
Use Pickle-based state persistence to keep track of containers.
Eventually, we'll switch to SQLite.

Checkpoint/restore support in Docker is experimental and weak. Podman
has better support for CRIU, but there are limitations in their HTTP
server implementation, thus we shell out to the podman CLI. A better
option would be to integrate with Podman's libpod but that would mean
writing stuff in Go, which is currently a no-go. A more extreme option
would be to call out to runc/criu directly. Later, perhaps.

Switch back to runc runtime, since gVisor (runsc) doesn't seem to work
with checkpoint/restore, or I can't figure out how to make it work.

Add basic integration tests.